# [store] Unique reports _before_ storing