--- a/tools/report-converter/codechecker_report_converter/report/hash.py
+++ b/tools/report-converter/codechecker_report_converter/report/hash.py
@@ -205,5 +205,5 @@ def get_report_path_hash(report: Report) -> str:
     if not report_path_hash:
         LOG.error('Failed to generate report path hash: %s', report)
 
-    LOG.debug(report_path_hash)
+    # LOG.debug(report_path_hash)
     return __str_to_hash(report_path_hash)